requirements.txt
google-cloud-bigquery==1.5.0
psycopg2==2.7.5